feat(player-search-form): migrates to material ui
Migrates component to use MaterialUI components instead of chakra-ui

re #144
